Version 1.0

+++ api/app/Http/Controllers/BankController.php
@@ -2,12 +2,22 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Account;
 use App\Models\Bank;
+use App\Models\Professor;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
 class BankController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware("permission:voir banque")->only(["index", "show"]);
+        $this->middleware("permission:modifier banque")->only(["update"]);
+        $this->middleware("permission:creer banque")->only(["store"]);
+        $this->middleware("permission:supprimer banque")->only(["destroy"]);
+    }
     public function index()
     {
         return Bank::all();
@@ -31,13 +41,25 @@ class BankController extends Controller
     public function update(Request $request, $id)
     {
         $request->validate(['name' => 'required', 'code' => 'required']);
-        return DB::table('banks')->whereId($id)->first()->update($request->all());
+        $bank = Bank::find($id);
+        $bank->name = $request->name;
+        $bank->code = $request->code;
+        $bank->save();
+        return $this->show($id);
     }
 
     public function destroy($id)
     {
-       return DB::table('banks')->whereId($id)->delete();
+        $professors = Account::whereBankId($id)->get();
+        if (count($professors) == 0)
+            return DB::table('banks')->whereId($id)->delete();
+        else return response()->json(["message" => "Cette bank est liée à un compte professeur."], 400);
     }
 
-
+    public function search($data)
+    {
+        return Bank::where('name', 'like', '%' . $data . '%')
+            ->orWhere('code', 'like', '%' . $data . '%')
+            ->get();
+    }
 }
